Move execute batch to normal transaction queue
The executeBatch function was starting a transaction but not blocking the transaction queue. This could cause a race condition where another transaction could be started from another thread and then sqlite with error out saying it cannot start a transaction inside a transaction. This was because normal `transaction` calls are behind a mutex on the JS side, but executeBatch was not part of this queue. By moving it inside we avoid nasty race conditions.